requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
psycopg2-binary
requests
dvc
apache-airflow-providers-postgres
pyarrow